Code refactore.
