+++ SpaceBall/SpaceBall/ConfigPaths.h
@@ -3,3 +3,5 @@
 const string WindowSettingsPath = "../Configs/Window Settings.ini";
 
 const string SupportedKeysPath = "../Configs/Supported keys.ini";
+
+const string ExceptionSavingDirectoryPath = "../Logs";
