+++ SpaceBall/SpaceBall/Debug/SpaceBall.log
@@ -1,5 +1,2 @@
-﻿  Game.cpp
-  Main.Cpp
-  RenderWindowBuilder.cpp
-  Generating Code...
+﻿  Main.Cpp
   SpaceBall.vcxproj -> C:\Users\AKijowski\source\repos\Space-Ball\SpaceBall\Debug\SpaceBall.exe
