--- SpaceBall/SpaceBall/Exception.cpp
@@ -1,13 +1,0 @@
-#include "stdafx.h"
-#include "Exception.h"
-
-Exception::Exception(string message)
-	: std::exception()
-{
-	_message = message;
-}
-
-string Exception::ToString()
-{
-	return _message;
-}
