--- SpaceBall/SpaceBall/Exception.h
@@ -1,12 +1,0 @@
-#pragma once
-
-class Exception : public std::exception
-{
-	private:
-		string _message;
-	
-	public:
-		Exception(string message);
-
-		virtual string ToString();
-};
