+++ SpaceBall/SpaceBall/IExceptionHandler.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "stdafx.h"
+#include "Exception.hpp"
+
+class IExceptionHandler
+{
+public:
+	virtual void HandleException(Exception* exception) = 0;
+};
