+++ SpaceBall/SpaceBall/Main.Cpp
@@ -1,5 +1,5 @@
 #include "stdafx.h"
-#include "Game.h"
+#include "Game.hpp"
 int main()
 {
 	Game game;
