+++ SpaceBall/SpaceBall/SpaceBall.vcxproj
@@ -131,26 +131,25 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Exception.cpp" />
-    <ClCompile Include="Game.cpp" />
     <ClCompile Include="Main.Cpp" />
-    <ClCompile Include="RenderWindowBuilder.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="SupportedKeysMapBuilder.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ConfigPaths.h" />
-    <ClInclude Include="Exception.h" />
-    <ClInclude Include="FailedWindowInitializationException.h" />
-    <ClInclude Include="FileNotFoundException.h" />
-    <ClInclude Include="Game.h" />
-    <ClInclude Include="RenderWindowBuilder.h" />
+    <ClInclude Include="Exception.hpp" />
+    <ClInclude Include="ExceptionHandler.hpp" />
+    <ClInclude Include="ExceptionHandlerBuilder.hpp" />
+    <ClInclude Include="FailedWindowInitializationException.hpp" />
+    <ClInclude Include="FileNotFoundException.hpp" />
+    <ClInclude Include="Game.hpp" />
+    <ClInclude Include="IExceptionHandler.hpp" />
+    <ClInclude Include="RenderWindowBuilder.hpp" />
     <ClInclude Include="stdafx.h" />
-    <ClInclude Include="SupportedKeysMapBuilder.h" />
-    <ClInclude Include="WrongFileContentException.h" />
+    <ClInclude Include="SupportedKeysMapBuilder.hpp" />
+    <ClInclude Include="WrongFileContentException.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
