--- SpaceBall/SpaceBall/SupportedKeysMapBuilder.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "FileNotFoundException.h"
-#include "WrongFileContentException.h"
-
-class SupportedKeysMapBuilder
-{
-	public:
-		SupportedKeysMapBuilder() {};
-
-		Map<string, int>* BuildSupportedKeysMap();
-};
