+++ SpaceBall/SpaceBall/WrongFileContentException.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include "stdafx.h"
+#include "Exception.hpp"
+
+class WrongFileContentException : public Exception
+{
+public:
+	WrongFileContentException(string filePath)
+		: Exception("File has wrong content. Path was :" + filePath)
+	{
+	}
+};
